CHROMIUM: linux-firmware: Add symlink for Intel audio firmware
In the upstream linux kernel, as well as the linux-firmware git tree,
the intel audio firmware is referred with the name
fw_sst_0f28.bin-48kHz_i2s_master" (note the 48kHz in name).

Upstream linux-firmware:
7551a3a fw_sst_0f28: Add firmware for Intel Baytrail SST DSP

Upstream linux kernel:
9b11eb4 ASoC: Intel: Update Baytrail ADSP firmware name

Yet, in our tree, this firmware is named "fw_sst_0f28.bin-i2s_master"
(no 48kHz in name). Hence when I load 3.18 on baytrail systems, the
kernel searches for, and fails to find the *48kHz* firmware file:
sst-acpi 80860F28:00: Direct firmware load for intel/fw_sst_0f28.bin-48kHz_i2s_master failed with error -2

I was tempted to cherry-pick the upstream commit
so that we'll have new upstream firmware image that will also have the
48kHz in name, however there are 2 issues with that:

1) Our fw image seems different than upstream (different size) and we
seem to have put fixes for suspend / resume in our image. I'm not sure
if the upstream version has those fixes and if we can use that.

2) We have a bunch of devices that still run older kernels (3.8/3.10)
that still refer to the file as fw_sst_0f28.bin-i2s_master (no 48kHz)

This taking the simplest route so that we do not disturb any oher
component - I made a symlink that points to the old firmware image - so
everyone is happy.

BUG=b:28634526
TEST=On glimmer, verified that the firmware gets loaded and "aplay -l"
     shows the audio device.

Change-Id: I190a9112563fdc5ff40fa9f4ab2e03b9b7d7fa3c
Reviewed-on: https://chromium-review.googlesource.com/349343
Commit-Ready: Benson Leung <bleung@chromium.org>
Tested-by: Benson Leung <bleung@chromium.org>
Reviewed-by: Benson Leung <bleung@chromium.org>