Official PCIe-USB Marvell 8997 16.68.1.p88 FW Release for Kevin
For chrome-os-partner:57914
One of the FW parameters related to IEEE-PS mode was getting an
incorrect value set after resume. Hence, FW goes to sleep more often
than expected and leads to Throughput degradation. This FW will ensure
that the values are correct before and after resume.

For chrome-os-partner:57319
Issue has been fixed in the BT-USB Open Source Driver and
/patch/9382113/ has been submitted upstream. This FW contains a
workaround to address the issue while waiting for the patch to be
accepted.

BUG=chrome-os-partner:57914, chrome-os-partner:57319
TEST=See bug

Change-Id: Idafe5ba195faee4ea4c7fbb82c0e93e0794430d3
Signed-off-by: Prashanth Ranganathan <rprashk@marvell.com>
Reviewed-on: https://chromium-review.googlesource.com/401703
Commit-Ready: Brian Norris <briannorris@chromium.org>
Tested-by: Brian Norris <briannorris@chromium.org>
Reviewed-by: Wei-Ning Huang <wnhuang@chromium.org>