CHROMIUM: ath10k: add WCN3990 firmware-5.bin
This is not actually a firmware; it just a firmware flags file,
containing a few flags and options to describe how the firmware should
work. Particularly, it contains the NON_BMI flag, which tells a driver
that the WLAN firmware will not be loaded via traditional host-initiated
methods, but it will instead be loaded "out of band," while we simply
wait for a "firmware ready" signal.

Generated by:

https://github.com/qca/qca-swiss-army-knife.git

  $ tools/scripts/ath10k/ath10k-fwencoder \
    --create --set-wmi-op-version=tlv --set-htt-op-version=tlv \
    --set-fw-api=5 --features=mgmt-tx-by-ref,non-bmi \
    --timestamp=1530303384

The fwencoder tool was recently updated to include the 'mgmt-txt-by-ref'
and 'non-bmi' flags. I used the '--timestamp' flag to match the firmware
I pieced together myself previously -- the only diff is in the the
length/padding of the FW_FEATURES TLV, and in the ordering of the TLVs
within the file.

This (along with plenty of other missing pieces that are still in
flight) was used to connect to a local AP.

BUG=b:78530936
TEST=it's complicated

Change-Id: I3c8d091411b58dcb0ed00b9f2c43466f83098e9c
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1125412
Commit-Ready: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>