CHROMIUM: add board binary file in WCN6855
ath11k driver has a dependency on board.bin file. This
patch provides a default board.bin file to enable Wifi.
Once Qualcomm publishes the updated board.bin file to
upstream, this change will be replaced with the upstream
board.bin. The upstreaming of board.bin is tracked in the
ticket b/197348036.

BUG=b:197348036
TEST=The board-2.bin installed in /lib/firmware

Signed-off-by: Abhishek Kumar <kuabhs@chromium.org>
Change-Id: I5eb3156d22509beab8c5753f274fb0af10c3e563
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/linux-firmware/+/3111666
Reviewed-by: Douglas Anderson <dianders@chromium.org>