iwlwifi: add core72-55 firmwares for So
Add firmware.

Build number: Core_build_core72-55
Revision number: 5d560e6d (So)

This is a modified version of the change chromium:3806864:
The changes in the iwlwifi-so-a0-gf-a0.pnvm are skipped here since this
file is shared with the board Brya and we don't want to introduce any
changes to Brya's firmware.

BUG=b:241121006
TEST=Cq

Cq-Depend: chromium:3844793
Change-Id: I45b64b0eee3edb60996294825b8fd7a9c6bf82bd
Signed-off-by: Gregory Greenman <gregory.greenman@intel.com>
Signed-off-by: Arowa Suliman <arowa@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/linux-firmware/+/3844842
Reviewed-by: Abhishek Kumar <kuabhs@chromium.org>
Reviewed-by: Nicolas Norvez <norvez@chromium.org>