BACKPORT: iwlwifi: add new FWs from core72-129 release
Add the -74.ucode firmwares for the currently supported hardware and
update the PNVM binaries.

Build number: Core_build_core72-129
Revision: a5e9588b (Qu, QuZ, cc, so, ty)

Signed-off-by: Gregory Greenman <gregory.greenman@intel.com>
(cherry picked from commit 98b55774bcc6c901cc985e8fab1404219a4ef196)

BUG=b:268140147
TEST=none

This patch adds a firmware file that is needed by the iwlwifi driver
for the kernel provided by the kernel-upstream USE flag. The proximal
failure is emitted by chromite/bin/test_image when creating an image:
```
The driver supports the following "iwlwifi-so-a0-hr" files: iwlwifi-so-a0-hr-b0-74.ucode
The rootfs provides the following "iwlwifi-so-a0-hr" files: iwlwifi-so-a0-hr-b0-81.ucode
[... snip ...]
Driver/firmware mismatch for iwlwifi-so-a0-hr
```
This is in support of bringing up a nissa-kernelnext builder that we
can use for continuous rebase testing.

When backporting this, I only took the file iwlwifi-so-a0-hr-b0-74.ucode
from the upstream commit, and fixed up WHENCE to only mention the
addition of that file.

Signed-off-by: Eizan Miyamoto <eizan@chromium.org>
Change-Id: I1c7c5a99233780ca801a48782c8988e0545a8b00
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/linux-firmware/+/4506034
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Nicolas Norvez <norvez@chromium.org>
Reviewed-by: Abhishek Kumar <kuabhs@chromium.org>